Nutrition service user target message

+++ NutritionService/Features/MealDetails/MealDetailsModelView.cs
@@ -13,9 +13,9 @@ namespace NutritionService.Features.MealDetails
             public string ImageUrl { get; set; } = string.Empty;
             public string VideoUrl { get; set; } = string.Empty;
             public int Calories { get; set; }
-            public decimal Protein { get; set; }
-            public decimal Carbohydrates { get; set; }
-            public decimal Fat { get; set; }
+            public double Protein { get; set; }
+            public double Carbohydrates { get; set; }
+            public double Fat { get; set; }
             public string DifficultyLevel { get; set; } = string.Empty;
             public bool isPremium { get; set; } = false;
             public List<IngredientModelView> Ingredients { get; set; } = new List<IngredientModelView>();
